+++ iosHyperskillApp/.swiftlint.yml
@@ -132,7 +132,6 @@ only_rules:
   - switch_case_alignment
   - switch_case_on_newline
   - syntactic_sugar
-  - superfluous_disable_command
   - toggle_bool
   - trailing_comma
   - trailing_newline
