+++ iosHyperskillApp/iosHyperskillApp/Sources/Views/Skeletons/SkeletonRoundedButton.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+struct SkeletonRoundedButton: View {
+    let height: CGFloat = 44
+
+    var body: some View {
+        SkeletonRoundedView()
+            .frame(height: height)
+    }
+}
+
+struct SkeletonRoundedButton_Previews: PreviewProvider {
+    static var previews: some View {
+        SkeletonRoundedButton()
+    }
+}
